requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
python-magic-bin==0.4.14
aiofiles==23.2.1
numpy==1.26.2